Harden ADE runtime connectivity after dual-brain and auth races.
Stop daily Clerk invalid_grant sign-outs with a live-peer refresh journal mutex, tell the truth when this computer is missing from the account directory, stop blaming sshd for refused ADE ports, reap same-channel wedged sync holders, prefer and hot-migrate back to 8787, and cap the chat event-history ring that was wedging the brain under a 1GB heap.
