+++ apps/ade-cli/src/services/account/accountAuthService.ts
@@ -923,6 +923,11 @@ export function createAccountAuthService(args: {
   sessionMutationSource?: AccountSessionMutationSource | null;
   /** Overrides `process.pid` on audit log lines and journal entries. */
   pid?: number;
+  /**
+   * Whether a journaled pid is still running. Tests inject this so a fixture
+   * pid cannot be mistaken for a live peer on the host.
+   */
+  pidAlive?: (pid: number) => boolean;
 }): AccountAuthService {
   const fetchImpl = args.fetchImpl ?? ((input, init) => fetch(input, init));
   const now = args.now ?? Date.now;
@@ -1106,6 +1111,7 @@ export function createAccountAuthService(args: {
     pid: mutationPid,
     source: mutationSource,
     log: logSessionMutation,
+    ...(args.pidAlive ? { pidAlive: args.pidAlive } : {}),
   });
 
   const persistSession = (
@@ -2275,149 +2281,176 @@ export function createAccountAuthService(args: {
         };
         let token: TokenResponse | null = null;
         let config: AccountOAuthConfig | null = null;
-        for (let attempt = 0; attempt < 2; attempt += 1) {
-          const refreshRecord = refreshSnapshot.session;
-          config = refreshRecord.oauthConfig
-            ? normalizeOAuthConfig(refreshRecord.oauthConfig)
-            : await resolveOAuthConfig();
-          const tokenGeneration = accountTokenGeneration(refreshRecord.refreshToken) ?? "";
-          // Read the journal BEFORE writing ours: an entry still naming this
-          // exact token generation means some process already started an
-          // exchange against it and never finished. The `invalid_grant` that
-          // follows is then explainable by that interruption, not proof that
-          // the grant is dead, so it must not condemn the session.
-          const priorJournal = rotationJournal.read();
-          const interruptedRotation = priorJournal != null
-            && priorJournal.oldRefreshTokenHash === tokenGeneration;
-          if (interruptedRotation) {
-            logSessionMutation({
-              action: "rotation_journal_interrupted",
-              reason: "unfinished_rotation_observed",
-              level: "warn",
-              tokenGeneration,
-              outcome: `started_at:${priorJournal.startedAt} pid:${priorJournal.pid} source:${priorJournal.source ?? "unknown"}`,
+        // Generation we journaled for this exchange. Cleared in `finally` so a
+        // network/timeout/early-return path cannot leave a live-pid journal that
+        // peers wait on forever (the sticky mutex that looks like daily logout).
+        let journaledGeneration: string | null = null;
+        const accessTokenStillFresh = (session: AccountSessionRecord): boolean => {
+          const expiresAtMs = Date.parse(
+            accessTokenExpiresAt(session.accessToken) ?? session.expiresAt,
+          );
+          return Number.isFinite(expiresAtMs) && expiresAtMs > now() + ACCESS_TOKEN_REFRESH_SKEW_MS;
+        };
+        try {
+          for (let attempt = 0; attempt < 2; attempt += 1) {
+            const refreshRecord = refreshSnapshot.session;
+            config = refreshRecord.oauthConfig
+              ? normalizeOAuthConfig(refreshRecord.oauthConfig)
+              : await resolveOAuthConfig();
+            const tokenGeneration = accountTokenGeneration(refreshRecord.refreshToken) ?? "";
+            // Compare-and-swap the journal before talking to Clerk. A live peer
+            // already exchanging this grant must be waited out: Clerk refresh
+            // tokens are single-use, and a second POST is what produced the
+            // daily `invalid_grant` / mark_dead sign-outs. A dead peer's journal
+            // is taken over and treated as an interrupted rotation, so the
+            // `invalid_grant` that follows is not definitive.
+            const begin = rotationJournal.tryBegin({
+              oldRefreshTokenHash: tokenGeneration,
+              userId: refreshRecord.userId,
             });
-          }
-          rotationJournal.write({
-            oldRefreshTokenHash: tokenGeneration,
-            userId: refreshRecord.userId,
-          });
-          try {
-            token = await postTokenForm({
-              fetchImpl,
-              tokenUrl: `${config.issuer}/oauth/token`,
-              signal: sharedSignal,
-              body: {
-                grant_type: "refresh_token",
-                refresh_token: refreshRecord.refreshToken!,
-                client_id: config.clientId,
-              },
-            });
-            break;
-          } catch (error) {
-            if (
-              !(error instanceof AccountTokenRequestError)
-              || error.oauthErrorCode !== "invalid_grant"
-            ) {
-              throw error;
+            if (begin.kind === "peer_in_flight") {
+              const rotation = await waitForRefreshRotation(refreshSnapshot, sharedSignal);
+              if (rotation.kind === "rotated") return rotation.snapshot.session;
+              if (rotation.kind !== "unchanged") return null;
+              throw new Error(
+                "ADE is refreshing your account sign-in in another process. Retry in a moment.",
+              );
             }
-            // The desktop and brain share this credential. A peer that won a
-            // rotating refresh exchange may not have persisted its replacement
-            // by the time Clerk rejects our old token, so poll before declaring
-            // the grant dead. The window out-waits the credential store's lock
-            // timeout, so a winner still queued for the lock cannot lose.
-            let rotation = await waitForRefreshRotation(refreshSnapshot, sharedSignal);
-            if (rotation.kind === "rotated" && attempt === 0) {
-              refreshSnapshot = rotation.snapshot;
-              continue;
-            }
-            if (rotation.kind !== "unchanged") return null;
-            if (interruptedRotation) {
-              // An interrupted journal makes this rejection ambiguous: the
-              // stored token may already have been spent by the process that
-              // died. Spend one more rotation-wait cycle, then give up for this
-              // attempt WITHOUT condemning the session. Clearing the journal
-              // makes the next refresh definitive, so an actually-dead grant
-              // still reaches the needs-re-auth state one attempt later.
-              rotation = await waitForRefreshRotation(refreshSnapshot, sharedSignal);
+            journaledGeneration = tokenGeneration;
+            const interruptedRotation = begin.takeover;
+            try {
+              token = await postTokenForm({
+                fetchImpl,
+                tokenUrl: `${config.issuer}/oauth/token`,
+                signal: sharedSignal,
+                body: {
+                  grant_type: "refresh_token",
+                  refresh_token: refreshRecord.refreshToken!,
+                  client_id: config.clientId,
+                },
+              });
+              break;
+            } catch (error) {
+              if (
+                !(error instanceof AccountTokenRequestError)
+                || error.oauthErrorCode !== "invalid_grant"
+              ) {
+                throw error;
+              }
+              // The desktop and brain share this credential. A peer that won a
+              // rotating refresh exchange may not have persisted its replacement
+              // by the time Clerk rejects our old token, so poll before declaring
+              // the grant dead. The window out-waits the credential store's lock
+              // timeout, so a winner still queued for the lock cannot lose.
+              let rotation = await waitForRefreshRotation(refreshSnapshot, sharedSignal);
               if (rotation.kind === "rotated" && attempt === 0) {
+                // A peer already persisted a usable pair. Serving it avoids a
+                // second Clerk POST that would burn their new rotating grant.
+                if (accessTokenStillFresh(rotation.snapshot.session)) {
+                  return rotation.snapshot.session;
+                }
                 refreshSnapshot = rotation.snapshot;
                 continue;
               }
               if (rotation.kind !== "unchanged") return null;
-              rotationJournal.clear("interrupted_rotation_inconclusive", tokenGeneration);
-              logSessionMutation({
-                action: "rotation_journal_interrupted",
-                reason: "invalid_grant_not_definitive",
-                level: "warn",
-                oauthErrorCode: error.oauthErrorCode,
-                tokenGeneration,
-                outcome: "session_preserved",
-              });
+              if (interruptedRotation) {
+                // An interrupted journal makes this rejection ambiguous: the
+                // stored token may already have been spent by the process that
+                // died. Spend one more rotation-wait cycle, then give up for this
+                // attempt WITHOUT condemning the session. Clearing the journal
+                // makes the next refresh definitive, so an actually-dead grant
+                // still reaches the needs-re-auth state one attempt later.
+                rotation = await waitForRefreshRotation(refreshSnapshot, sharedSignal);
+                if (rotation.kind === "rotated" && attempt === 0) {
+                  if (accessTokenStillFresh(rotation.snapshot.session)) {
+                    return rotation.snapshot.session;
+                  }
+                  refreshSnapshot = rotation.snapshot;
+                  continue;
+                }
+                if (rotation.kind !== "unchanged") return null;
+                rotationJournal.clear("interrupted_rotation_inconclusive", tokenGeneration);
+                journaledGeneration = null;
+                logSessionMutation({
+                  action: "rotation_journal_interrupted",
+                  reason: "invalid_grant_not_definitive",
+                  level: "warn",
+                  oauthErrorCode: error.oauthErrorCode,
+                  tokenGeneration,
+                  outcome: "session_preserved",
+                });
+                throw error;
+              }
+              const marked = markStoredSessionRejectedIfExact(
+                refreshSnapshot.raw,
+                refreshSnapshot.session,
+                error.oauthErrorCode,
+              );
+              if (!marked && readSessionSnapshot().raw !== refreshSnapshot.raw) {
+                return null;
+              }
               throw error;
             }
-            const marked = markStoredSessionRejectedIfExact(
-              refreshSnapshot.raw,
-              refreshSnapshot.session,
-              error.oauthErrorCode,
-            );
-            if (!marked && readSessionSnapshot().raw !== refreshSnapshot.raw) {
-              return null;
-            }
-            throw error;
           }
-        }
-        if (!token || !config) {
-          throw new Error("ADE account session expired. Run `ade login` again.");
-        }
-        if (authEpoch !== epochAtJoin) return null;
-        const obtainedAtMs = now();
-        const refreshed = await buildSessionRecord(
-          token,
-          refreshSnapshot.session,
-          undefined,
-          config,
-          { fetchUserinfo: false, obtainedAtMs, signal: sharedSignal },
-        );
-        if (authEpoch !== epochAtJoin) return null;
-        if (!persistRefreshedSessionIfCurrent(
-          refreshed,
-          refreshSnapshot.raw,
-          "refresh_token_rotated",
-          accountTokenGeneration(refreshSnapshot.session.refreshToken),
-        )) {
-          return null;
-        }
-
-        // The rotated access/refresh pair is durable before optional profile
-        // enrichment. Identity is carried from the previously verified subject,
-        // so avoidable userinfo latency cannot expose a stale refresh token to a
-        // second process.
-        let enriched: AccountSessionRecord;
-        try {
-          enriched = await buildSessionRecord(
+          if (!token || !config) {
+            throw new Error("ADE account session expired. Run `ade login` again.");
+          }
+          if (authEpoch !== epochAtJoin) return null;
+          const obtainedAtMs = now();
+          const refreshed = await buildSessionRecord(
             token,
             refreshSnapshot.session,
             undefined,
             config,
-            { obtainedAtMs, signal: sharedSignal },
+            { fetchUserinfo: false, obtainedAtMs, signal: sharedSignal },
           );
-        } catch (error) {
-          if (sharedSignal.aborted) throw error;
-          // The rotated credential and verified prior subject are already
-          // durable. Optional profile enrichment must not make that successful
-          // refresh unusable.
-          return readSession() ?? refreshed;
+          if (authEpoch !== epochAtJoin) return null;
+          if (!persistRefreshedSessionIfCurrent(
+            refreshed,
+            refreshSnapshot.raw,
+            "refresh_token_rotated",
+            accountTokenGeneration(refreshSnapshot.session.refreshToken),
+          )) {
+            // Persist path clears our journaled generation; skip the finally clear.
+            journaledGeneration = null;
+            return null;
+          }
+          journaledGeneration = null;
+
+          // The rotated access/refresh pair is durable before optional profile
+          // enrichment. Identity is carried from the previously verified subject,
+          // so avoidable userinfo latency cannot expose a stale refresh token to a
+          // second process.
+          let enriched: AccountSessionRecord;
+          try {
+            enriched = await buildSessionRecord(
+              token,
+              refreshSnapshot.session,
+              undefined,
+              config,
+              { obtainedAtMs, signal: sharedSignal },
+            );
+          } catch (error) {
+            if (sharedSignal.aborted) throw error;
+            // The rotated credential and verified prior subject are already
+            // durable. Optional profile enrichment must not make that successful
+            // refresh unusable.
+            return readSession() ?? refreshed;
+          }
+          if (authEpoch !== epochAtJoin) return null;
+          const refreshedRaw = JSON.stringify(refreshed);
+          return persistRefreshedSessionIfCurrent(
+            enriched,
+            refreshedRaw,
+            "refresh_profile_enriched",
+          )
+            ? enriched
+            : readSession();
+        } finally {
+          if (journaledGeneration) {
+            rotationJournal.clear("refresh_finished", journaledGeneration);
+          }
         }
-        if (authEpoch !== epochAtJoin) return null;
-        const refreshedRaw = JSON.stringify(refreshed);
-        return persistRefreshedSessionIfCurrent(
-          enriched,
-          refreshedRaw,
-          "refresh_profile_enriched",
-        )
-          ? enriched
-          : readSession();
       })().finally(() => {
         clearTimeout(sharedRefreshTimer);
         refreshInFlight = null;
