+++ apps/ade-cli/src/services/account/accountSessionRotationJournal.ts
@@ -10,6 +10,11 @@
  * specific token generation, and is cleared once the replacement is durable. An
  * entry that survives means "the stored token may already have been consumed" —
  * the one `invalid_grant` that follows is not definitive.
+ *
+ * The same entry is also a mutex. Clerk refresh tokens are single-use: two
+ * live processes that both POST `/oauth/token` against one generation will
+ * make the loser look signed out. `tryBegin` compare-and-swaps the journal so
+ * a live peer's in-flight exchange is waited out, not raced.
  */
 
 import type { SyncCredentialStore } from "../credentials/credentialStore";
@@ -34,6 +39,10 @@ export type RotationJournalEntry = {
   source: AccountSessionMutationSource | null;
   userId: string | null;
 };
+
+export type RotationJournalBeginResult =
+  | { kind: "acquired"; takeover: boolean }
+  | { kind: "peer_in_flight"; entry: RotationJournalEntry };
 
 function asRecord(value: unknown): Record<string, unknown> {
   return value && typeof value === "object" && !Array.isArray(value)
@@ -67,6 +76,22 @@ export function parseRotationJournal(
   }
 }
 
+function describeJournalPeer(entry: RotationJournalEntry): string {
+  return `started_at:${entry.startedAt} pid:${entry.pid} source:${entry.source ?? "unknown"}`;
+}
+
+function defaultPidAlive(pid: number): boolean {
+  if (!Number.isFinite(pid) || pid <= 0) return false;
+  try {
+    process.kill(pid, 0);
+    return true;
+  } catch (error) {
+    // EPERM: the process exists; this caller just cannot signal it. Treat it
+    // as live so we do not steal a refresh that is still in flight.
+    return (error as NodeJS.ErrnoException).code === "EPERM";
+  }
+}
+
 export type RotationJournalArgs = {
   credentialStore: SyncCredentialStore;
   /** Clock, so tests can pin `startedAt`. */
@@ -79,12 +104,28 @@ export type RotationJournalArgs = {
     action: AccountSessionMutationAction;
     reason: string;
     tokenGeneration?: string | null;
+    level?: "info" | "warn";
+    outcome?: string;
   }) => void;
+  /**
+   * Whether a journaled pid is still running. Tests inject this so a fixture
+   * pid cannot be mistaken for a live peer on the host.
+   */
+  pidAlive?: (pid: number) => boolean;
 };
 
 export type RotationJournal = {
   read: () => RotationJournalEntry | null;
-  write: (entry: { oldRefreshTokenHash: string; userId: string | null }) => void;
+  /**
+   * Compare-and-swap begin. A live peer's journal for this (or any) generation
+   * is left untouched and reported as `peer_in_flight` so the caller waits
+   * instead of burning the rotating grant at Clerk. A dead peer's journal is
+   * taken over and treated as an interrupted rotation.
+   */
+  tryBegin: (entry: {
+    oldRefreshTokenHash: string;
+    userId: string | null;
+  }) => RotationJournalBeginResult;
   /**
    * `expectedTokenGeneration` scopes the clear to OUR entry. A peer may have
    * started its own rotation against a newer generation while ours was in
@@ -95,7 +136,14 @@ export type RotationJournal = {
   clear: (reason: string, expectedTokenGeneration?: string | null) => void;
 };
 
+type BeginDecision =
+  | { kind: "write"; takeover: boolean }
+  | { kind: "already_ours" }
+  | { kind: "peer_in_flight"; entry: RotationJournalEntry };
+
 export function createRotationJournal(args: RotationJournalArgs): RotationJournal {
+  const pidAlive = args.pidAlive ?? defaultPidAlive;
+
   const read = (): RotationJournalEntry | null => {
     try {
       return parseRotationJournal(
@@ -108,27 +156,113 @@ export function createRotationJournal(args: RotationJournalArgs): RotationJourna
     }
   };
 
-  const write = (entry: { oldRefreshTokenHash: string; userId: string | null }): void => {
+  const serialize = (entry: { oldRefreshTokenHash: string; userId: string | null }): string =>
+    JSON.stringify({
+      version: 1,
+      oldRefreshTokenHash: entry.oldRefreshTokenHash,
+      startedAt: new Date(args.now()).toISOString(),
+      pid: args.pid,
+      source: args.source,
+      userId: entry.userId,
+    });
+
+  const decide = (
+    existing: RotationJournalEntry | null,
+    entry: { oldRefreshTokenHash: string; userId: string | null },
+  ): BeginDecision => {
+    if (!existing) return { kind: "write", takeover: false };
+    if (existing.pid === args.pid && existing.oldRefreshTokenHash === entry.oldRefreshTokenHash) {
+      return { kind: "already_ours" };
+    }
+    if (existing.pid !== args.pid && pidAlive(existing.pid)) {
+      return { kind: "peer_in_flight", entry: existing };
+    }
+    return {
+      kind: "write",
+      takeover: existing.oldRefreshTokenHash === entry.oldRefreshTokenHash
+        && existing.pid !== args.pid,
+    };
+  };
+
+  const persistBegin = (entry: { oldRefreshTokenHash: string; userId: string | null }): void => {
     try {
       args.credentialStore.setSync(
         ACCOUNT_SESSION_ROTATION_JOURNAL_KEY,
-        JSON.stringify({
-          version: 1,
-          oldRefreshTokenHash: entry.oldRefreshTokenHash,
-          startedAt: new Date(args.now()).toISOString(),
-          pid: args.pid,
-          source: args.source,
-          userId: entry.userId,
-        }),
+        serialize(entry),
       );
-      args.log({
-        action: "rotation_journal_begin",
-        reason: "refresh_exchange_started",
-        tokenGeneration: entry.oldRefreshTokenHash,
-      });
     } catch {
       // Best effort: without a journal this refresh simply behaves the way it
       // did before, so a store write failure must not block the exchange.
+    }
+  };
+
+  const finalizeBegin = (
+    decision: BeginDecision,
+    entry: { oldRefreshTokenHash: string; userId: string | null },
+    options: { alreadyPersisted?: boolean } = {},
+  ): RotationJournalBeginResult => {
+    if (decision.kind === "peer_in_flight") {
+      args.log({
+        action: "rotation_journal_interrupted",
+        reason: "peer_refresh_in_flight",
+        tokenGeneration: entry.oldRefreshTokenHash,
+        level: "warn",
+        outcome: describeJournalPeer(decision.entry),
+      });
+      return decision;
+    }
+    if (decision.kind === "already_ours") {
+      return { kind: "acquired", takeover: false };
+    }
+    if (decision.takeover) {
+      args.log({
+        action: "rotation_journal_interrupted",
+        reason: "dead_peer_journal_taken_over",
+        tokenGeneration: entry.oldRefreshTokenHash,
+        level: "warn",
+      });
+    }
+    if (!options.alreadyPersisted) {
+      persistBegin(entry);
+    }
+    args.log({
+      action: "rotation_journal_begin",
+      reason: "refresh_exchange_started",
+      tokenGeneration: entry.oldRefreshTokenHash,
+    });
+    return { kind: "acquired", takeover: decision.takeover };
+  };
+
+  const tryBegin = (entry: {
+    oldRefreshTokenHash: string;
+    userId: string | null;
+  }): RotationJournalBeginResult => {
+    try {
+      const updateSync = args.credentialStore.updateSync;
+      if (!updateSync) {
+        return finalizeBegin(decide(read(), entry), entry);
+      }
+
+      let decision: BeginDecision | undefined;
+      updateSync.call(args.credentialStore, (values) => {
+        const existing = parseRotationJournal(values[ACCOUNT_SESSION_ROTATION_JOURNAL_KEY]);
+        decision = decide(existing, entry);
+        if (decision.kind === "peer_in_flight" || decision.kind === "already_ours") {
+          return false;
+        }
+        values[ACCOUNT_SESSION_ROTATION_JOURNAL_KEY] = serialize(entry);
+        return true;
+      });
+      if (!decision) {
+        // The store declined to run the updater. Proceed without a journal
+        // rather than blocking the exchange.
+        return { kind: "acquired", takeover: false };
+      }
+      return finalizeBegin(decision, entry, {
+        alreadyPersisted: decision.kind === "write",
+      });
+    } catch {
+      return { kind: "acquired", takeover: false };
     }
   };
 
@@ -155,5 +289,5 @@ export function createRotationJournal(args: RotationJournalArgs): RotationJourna
     }
   };
 
-  return { read, write, clear };
+  return { read, tryBegin, clear };
 }
