+++ apps/desktop/src/renderer/components/remoteTargets/remoteMachineModel.ts
@@ -320,9 +320,12 @@ export function formatRemoteTargetError(error: unknown): string {
     .replace(/^Error invoking remote method '[^']+':\s*/i, "")
     .replace(/^Error:\s*/i, "")
     .trim();
+  const sshHint = /\b(?:ssh|sshd|remote login)\b/i.test(message);
 
-  if (/^(?:read\s+)?ECONNRESET$/i.test(message)) {
-    return "SSH server closed the connection before ADE could finish the SSH handshake. Check that Remote Login/sshd is enabled on the remote machine and try again.";
+  if (/ECONNRESET/i.test(message)) {
+    return sshHint
+      ? "SSH server closed the connection before ADE could finish the SSH handshake. Check that Remote Login/sshd is enabled on the remote machine and try again."
+      : "The connection was reset before ADE could finish connecting. Check that the machine is awake and reachable, then try again.";
   }
 
   if (
@@ -342,11 +345,15 @@ export function formatRemoteTargetError(error: unknown): string {
       message,
     )
   ) {
-    return "SSH did not finish connecting. Check that the machine is awake, reachable on Tailscale or LAN, and Remote Login is enabled.";
+    return sshHint
+      ? "SSH did not finish connecting. Check that the machine is awake, reachable on Tailscale or LAN, and Remote Login is enabled."
+      : "ADE did not finish connecting. Check that the machine is awake and reachable on Tailscale or LAN.";
   }
 
   if (/ECONNREFUSED/i.test(message)) {
-    return "The machine refused the SSH connection. Check the port and make sure Remote Login/sshd is running.";
+    return sshHint
+      ? "The machine refused the SSH connection. Check the port and make sure Remote Login/sshd is running."
+      : "The machine refused the connection. Check that ADE is running on that computer and reachable on the saved port.";
   }
 
   if (
